+++ AtlasUtil/include/AtlasUtil/Annotate.h
@@ -2,6 +2,6 @@
 #include <llvm/IR/Metadata.h>
 #include <llvm/IR/Module.h>
 void Annotate(llvm::Module *M);
-void Annotate(llvm::Function *F, int &startingIndex);
+void Annotate(llvm::Function *F, uint64_t &startingIndex);
 int64_t GetBlockID(llvm::BasicBlock *BB);
-void SetBlockID(llvm::BasicBlock *BB, int64_t i);
+void SetBlockID(llvm::BasicBlock *BB, uint64_t i);
